requirements.txt
flask
robotframework-seleniumwire
selenium
tldextract